Slightly more permanent patch for street-sign.jpg
I've made a slightly more permanent patch for the street-sign.jpg file which just grabs the file from from the URL specified and places it in the resources/phase_4/maps folder without any of that additional round-about bullshit Disney implemented.

I've also adjusted the other URLs. Gamenews will fallback to the latest archive.org one while the main URL leads to the ASCDN. And since the CDN is still missing the files, it'll just redirect to the archive until I've got it all working.